requirements.txt
auraloss==0.4.0
causal_conv1d==1.4.0
einops==0.8.2
extensions==0.4
librosa==0.10.1
lightning==2.1.3
mamba_ssm==2.2.2
matplotlib==3.7.3
numpy==1.24.3
pykeops==2.3
pyloudnorm==0.1.1
pytorch_lightning==1.6.0
PyYAML==6.0.3
scipy==1.17.1
soundfile==0.12.1
thop==0.1.1.post2209072238
torch==2.1.1
torchaudio==2.1.1
torchsummary==1.5.1
tqdm==4.66.1
ultralytics_thop==2.0.13
